third_party/blink/web_tests/svg/dynamic-updates/SVGFEDiffuseLightingElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.in1.baseVal is "SourceGraphic"
PASS gradientElement.in1.baseVal is "blur"
PASS successfullyParsed is true

TEST COMPLETE
